% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{attributes}
\alias{attributes}
\title{Attribute List}
\format{A tibble with \Sexpr{nrow(seabirddietDB::attributes)} rows and \Sexpr{ncol(seabirddietDB::attributes)} variables}
\usage{
attributes
}
\description{
Detailed attribute list of columns in \code{seabirddiet}. For further details on 
attribute fields see relevant \href{https://ropensci.github.io/EML/articles/creating-EML.html#attribute-metadata}{documentation}
in package \pkg{EML}.
}
\keyword{datasets}
