% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{bind_taxonomy}
\alias{bind_taxonomy}
\title{Bind taxonomy columns to dataset}
\usage{
bind_taxonomy(data, taxonomy, role = c("prey", "pred"))
}
\arguments{
\item{data}{dataset}

\item{role}{whether the taxonomy relates to}
}
\description{
Bind taxonomy columns to dataset
}
