% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_fuzzy_values}
\alias{check_fuzzy_values}
\title{Fuzzy match check for typos}
\usage{
check_fuzzy_values(check_var, df, threshold = 0.1)
}
\arguments{
\item{check_var}{character string, variable name to check}

\item{df}{dataset}

\item{threshold}{distance threshold. Matches with distances greater than the
threshold are considered non-matches}
}
\value{
a tibble of potential typos
}
\description{
Fuzzy match check for typos
}
