% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{classification}
\alias{classification}
\title{Classification table}
\format{A tibble with \Sexpr{nrow(seabirddietDB::classification)} rows and \Sexpr{ncol(seabirddietDB::classification)} variables}
\usage{
classification
}
\description{
Taxonomic classification table of prey and predator taxa contained in \code{seabirddiet} columns \code{pred_species} 
and \code{prey_taxon}.
}
\keyword{datasets}
