% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{classify_validname}
\alias{classify_validname}
\title{Title}
\usage{
classify_validname(valid_name, ranks = c("phylum", "class", "order",
  "family", "genus", "species"))
}
\arguments{
\item{ranks}{}

\item{base_name}{}
}
\description{
Title
}
