% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{get_ref_tbl}
\alias{get_ref_tbl}
\title{Extract table of distinct references}
\usage{
get_ref_tbl(ref_tidy)
}
\arguments{
\item{ref_tidy}{tidy reference tibble}
}
\value{
a tibble containing the ref_id and valid reference fields of distinct
references.
}
\description{
Extract table of distinct references
}
\examples{
\dontrun{
get_ref_tbl(ref_tidy)
}
}
