% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_row_base}
\alias{get_row_base}
\title{Get row taxonomic base}
\usage{
get_row_base(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy data.frame}
}
\value{
the data.frame supplied with a column containing the taxonomic base
of each row appended.
}
\description{
Get the taxonomic base of each row in the original \code{taxonomy.tsv}.
}
