% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-recode.R
\name{mn_edit_df}
\alias{mn_edit_df}
\title{Launch interactive editable table}
\usage{
mn_edit_df(df)
}
\arguments{
\item{df}{data.frame to edit}
}
\value{
an html editable table widget of \code{df}
}
\description{
Launch interactive editable table
}
\examples{
\dontrun{
mn_edit_df(mtcars)
}
}
