% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-recode.R
\name{mn_recode}
\alias{mn_recode}
\title{Recode the values in the column of a data.frame using a manual recode table}
\usage{
mn_recode(df, recode_tbl, recode_target = NULL)
}
\arguments{
\item{df}{a data.frame in which column values are to be recoded.}

\item{recode_tbl}{a manual recoding table, as output from \code{mn_recode_tbl()}}

\item{recode_target}{character string. The name of the column to be recoded
if it differs from the first named column of \code{recode_tbl}}
}
\value{
the orgininal \code{df} with values in the \code{recode_target} column recode using
look up vales in \code{recode_tbl}
}
\description{
Recode the values in the column of a data.frame using a manual recode table
}
