% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual-recode.R
\name{mn_recode_tbl}
\alias{mn_recode_tbl}
\title{Launch interactive manual recode table}
\usage{
mn_recode_tbl(df, recode_var, pattern = NULL)
}
\arguments{
\item{df}{a data.frame or tibble}

\item{recode_var}{the unquoted name of the column for which recoding is required}

\item{pattern}{optional character string of regex pattern to match.
Matches are highlighted in the \code{pattern} column}
}
\value{
an html editable table widget of the selected variable and an editable
recode column.
}
\description{
Create and launch an editable interactive manual recoding look up table.
Edit entries and export to a variety of formats. Tables can be used to join manual
corrections or through functions like \code{plyr::revalue}
}
\examples{
\dontrun{
mn_recode_tbl(mtcars, cyl)
mn_recode_tbl(mtcars, cyl, "4")
}
}
