% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-tables.R
\name{rd_datatable}
\alias{rd_datatable}
\title{insert interactive table in html rd.}
\usage{
rd_datatable(df, width = "100\%", ...)
}
\arguments{
\item{df}{data.frame to}

\item{width}{width of data.frame}

\item{...}{additional arguments passed to datatable}
}
\description{
insert interactive table in html rd.
}
