% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{recode_spp_manual}
\alias{recode_spp_manual}
\title{Recode species with manual corrections table}
\usage{
recode_spp_manual(df, path = here::here("data-raw", "manual_corrections",
  "worrms_nomatches_verified.csv"), df_type = c("taxonomy", "data"))
}
\arguments{
\item{df}{data.frame in which to recode species}

\item{path}{the path to the manual match csv}

\item{df_type}{whether the data.frame supplied is the \code{taxonomy} df or the
seabirddiet \code{data} df.}
}
\value{
the data.frame supplied with the species names recoded
}
\description{
Recode species with manual corrections table
}
