% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{ref_bind_matched_missing_manual}
\alias{ref_bind_matched_missing_manual}
\title{Title}
\usage{
ref_bind_matched_missing_manual(seabirddiet, ref_manual)
}
\arguments{
\item{seabirddiet}{}

\item{ref_manual}{}
}
\description{
Title
}
