% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{ref_find_missing_manual}
\alias{ref_find_missing_manual}
\title{Find references missing from the ref_manual db}
\usage{
ref_find_missing_manual(data, ref_manual)
}
\arguments{
\item{data}{}

\item{ref_manual}{}
}
\description{
Find references missing from the ref_manual db
}
