% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{references}
\alias{references}
\title{Reference list}
\format{A tibble with \Sexpr{nrow(seabirddietDB::references)} rows and \Sexpr{ncol(seabirddietDB::references)} variables}
\usage{
references
}
\description{
Reference list linking the reference identifiers (\code{ref_id}) to the full reference 
they relate to (\code{ref_valid})
}
\keyword{datasets}
