% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sbd_filter}
\alias{sbd_filter}
\title{Filter Seabird Diet database}
\usage{
sbd_filter(db = seabirddietDB::seabirddiet, pred_species = NULL,
  prey_taxon = NULL, year = NULL, metrics = c("freq_occ",
  "freq_biomass", "freq_num"))
}
\arguments{
\item{db}{Seabird Diet database to filter. Defaults to `seabirddiet`}

\item{pred_species}{character vector. Predator species to include}

\item{prey_taxon}{character vector. Prey taxa to include}

\item{year}{integer vector. Year(s) to include}

\item{metrics}{which dietary metrics to include. Defaults to frequency of occurence (`freq_occ`)}
}
\value{
a filtered tibble of the supplied database
}
\description{
Filter Seabird Diet database
}
\examples{
sbd_filter(year = 2012)
sbd_filter(pred_species = "Alca torda", year = 2012)
}
