% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabirddietDB.R
\docType{package}
\name{seabirddietDB}
\alias{seabirddietDB}
\alias{seabirddietDB-package}
\title{\code{seabirddietDB} package}
\description{
Seabird Diet Database
}
\details{
See the README on
\href{https://github.com/annakrystalli/seabirddietDB/blob/master/README.md}{GitHub}
}
