% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{tidy_refs}
\alias{tidy_refs}
\title{Tidy references}
\usage{
tidy_refs(ref_manual)
}
\arguments{
\item{ref_manual}{dataframe of manual corrections, output in the reference
processing workbook.}
}
\value{
a long, tidy tibble of references
}
\description{
Separate and tidy references into a long format.
}
\examples{
\dontrun{
tidy_refs(ref_manual)
}
}
