% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{worrms_validate}
\alias{worrms_validate}
\title{Validate species in taxonomy against WoRMS database}
\usage{
worrms_validate(names)
}
\arguments{
\item{names}{vector of names to check}
}
\value{
the data.frame with the \code{aphia_id}, the valid aphia id and valid
name returned for the queried for the original species name
}
\description{
Validate species in taxonomy against WoRMS database
}
