% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{yr_to_isorange}
\alias{yr_to_isorange}
\title{Extract date ranges from a vector of years}
\usage{
yr_to_isorange(x, type = c("both", "start", "end"))
}
\arguments{
\item{x}{vector of years}

\item{type}{what output to return. \code{both} for the range,
\code{start} for only the start date, \code{end} for the end date}
}
\value{
years are converted to ISO dates of 1st Jan for a \code{begin} date
and 31st Dec for \code{end} dates. If \code{type = "range"} a vector of length 2
is returned. Otherwise a named character string of the specified
boundary is returned.
}
\description{
Extract date ranges from a vector of years
}
